Remove more unused code
